+++ test/framework/utils/Baseline.cpp
@@ -46,7 +46,7 @@ std::string DumpMD5(const void* bytes, size_t size) {
   unsigned char digest[CC_MD5_DIGEST_LENGTH] = {0};
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdeprecated-declarations"
-  CC_MD5(bytes, size, digest);
+  CC_MD5(bytes, static_cast<int>(size), digest);
 #pragma clang diagnostic pop
   char buffer[33];
   char* position = buffer;
